api/pyproject.toml
[project]
name = "octoxlabs-search-api"
version = "1.0.0"
description = "Octoxlabs Search API Case Study"
authors = [
    {name = "Burak Ece", email = "burakece9191@hotmail.com"}
]
readme = "README.md"
requires-python = "^3.11"
dependencies = [
    "django (>=5.0.2,<6.0.0)",
    "djangorestframework (>=3.14.0,<4.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "elasticsearch (>=8.12.1,<9.0.0)",
    "requests (>=2.31.0,<3.0.0)",
    "django-cors-headers (>=4.3.1,<5.0.0)",
    "pika (>=1.3.2,<2.0.0)",
    "drf-yasg (>=1.21.7,<2.0.0)",
    "psycopg2-binary (>=2.9.9,<3.0.0)"
]

[tool.poetry]
name = "octoxlabs-search-api"
version = "1.0.0"
description = "Octoxlabs Search API Case Study"
authors = ["Burak Ece <burakece9191@hotmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.2"
djangorestframework = "^3.14.0"
python-dotenv = "^1.0.1"
elasticsearch = "^8.12.1"
requests = "^2.31.0"
django-cors-headers = "^4.3.1"
pika = "^1.3.2"
drf-yasg = "^1.21.7"
psycopg2-binary = "^2.9.9"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.2"
pytest-django = "^4.8.0"
black = "^24.2.0"
flake8 = "^7.0.0"
isort = "^5.13.2"
coverage = "^7.4.3"
django-debug-toolbar = "^4.3.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
